Update dashboard for 7.5.1: remove create-workspace and use ready-to-go
Changes upstream mean that the create-workspace controller no longer
hosts devfile selection. Changes required to apply ephemeral mode by
default are instead moved to the ready-to-go-stacks controller.
