--- assembly/fabric8-ide-dashboard-war/src/app/workspaces/create-workspace/create-workspace.styl
@@ -1,44 +1,0 @@
-.create-workspace-progress
-  position absolute
-  top 56px
-  z-index 2
-
-.create-workspace-toolbar
-  .che-toolbar-header > div:last-child
-    bottom 9px
-
-.create-workspace-header-button .che-button-dropdown
-  button.che-button,
-  .fa-caret-down
-    min-height 36px
-    height 36px
-
-  .area-dropdown.dropdown-menu
-    top 36px
-
-
-md-content.create-workspace-content
-  background-color $very-light-grey-background-color
-  padding 15px
-
-  .che-label-container-label-name
-    text-transform uppercase
-    color $grey-input-label-color !important
-    font-weight bold !important
-
-  .che-label-container-content
-    min-width 850px
-
-  .create-workspace-content-hidden
-    visibility hidden
-
-  #create-workspace-footer-button
-    width 100%
-
-    button
-      margin 0 !important
-      width 100% !important
-
-  // rh-che specific changes
-  .ephemeral-mode-toggle
-    margin 0 !important
