+++ assembly/fabric8-ide-dashboard-war/src/app/workspaces/create-workspace/ready-to-go-stacks/ready-to-go-stacks.styl
@@ -1,0 +1,6 @@
+.ready-to-go-progress
+  z-index 2
+
+// rh-che specific changes
+.ephemeral-mode-toggle
+  margin 0 !important
